pyproject.toml
[project]
name = "outlook-timetable-extractor"
version = "0.1.0"
description = "A Python automation script that queries your Outlook mailbox, finds the latest timetable email, extracts your specific batch's timetable page from the PDF attachment, and saves it as an image — automatically."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "msal>=1.35.1",
    "pymupdf>=1.27.2.2",
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.7",
    "requests>=2.32.5",
]
